Move Caches to Simplepie namespace
