+++ library/SimplePie/Cache/Base.php
@@ -41,75 +41,15 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-/**
- * Base for cache objects
- *
- * Classes to be used with {@see SimplePie_Cache::register()} are expected
- * to implement this interface.
- *
- * @package SimplePie
- * @subpackage Caching
- */
-interface SimplePie_Cache_Base
-{
-	/**
-	 * Feed cache type
-	 *
-	 * @var string
-	 */
-	const TYPE_FEED = 'spc';
+use SimplePie\Cache\Base;
 
-	/**
-	 * Image cache type
-	 *
-	 * @var string
-	 */
-	const TYPE_IMAGE = 'spi';
+interface_exists('SimplePie\Cache\Base');
 
-	/**
-	 * Create a new cache object
-	 *
-	 * @param string $location Location string (from SimplePie::$cache_location)
-	 * @param string $name Unique ID for the cache
-	 * @param string $type Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
-	 */
-	public function __construct($location, $name, $type);
+// @trigger_error(sprintf('Using the "SimplePie_Cache_Base" class is deprecated since SimplePie version 1.x, use "SimplePie\Cache\Base" instead.'), \E_USER_DEPRECATED);
 
-	/**
-	 * Save data to the cache
-	 *
-	 * @param array|SimplePie $data Data to store in the cache. If passed a SimplePie object, only cache the $data property
-	 * @return bool Successfulness
-	 */
-	public function save($data);
-
-	/**
-	 * Retrieve the data saved to the cache
-	 *
-	 * @return array Data for SimplePie::$data
-	 */
-	public function load();
-
-	/**
-	 * Retrieve the last modified time for the cache
-	 *
-	 * @return int Timestamp
-	 */
-	public function mtime();
-
-	/**
-	 * Set the last modified time to the current time
-	 *
-	 * @return bool Success status
-	 */
-	public function touch();
-
-	/**
-	 * Remove the cache
-	 *
-	 * @return bool Success status
-	 */
-	public function unlink();
+if (\false) {
+	/** @deprecated since SimplePie 1.x, use "SimplePie\Cache\Base" instead */
+	interface SimplePie_Cache_Base extends Base
+	{
+	}
 }
-
-class_alias('SimplePie_Cache_Base', 'SimplePie\Cache\Base', false);
