+++ library/SimplePie/Cache/Memcached.php
@@ -41,128 +41,15 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-/**
- * Caches data to memcached
- *
- * Registered for URLs with the "memcached" protocol
- *
- * For example, `memcached://localhost:11211/?timeout=3600&prefix=sp_` will
- * connect to memcached on `localhost` on port 11211. All tables will be
- * prefixed with `sp_` and data will expire after 3600 seconds
- *
- * @package    SimplePie
- * @subpackage Caching
- * @author     Paul L. McNeely
- * @uses       Memcached
- */
-class SimplePie_Cache_Memcached implements SimplePie_Cache_Base
-{
-    /**
-     * Memcached instance
-     * @var Memcached
-     */
-    protected $cache;
+use SimplePie\Cache\Memcached;
 
-    /**
-     * Options
-     * @var array
-     */
-    protected $options;
+class_exists('SimplePie\Cache\Memcached');
 
-    /**
-     * Cache name
-     * @var string
-     */
-    protected $name;
+// @trigger_error(sprintf('Using the "SimplePie_Cache_Memcached" class is deprecated since SimplePie version 1.x, use "SimplePie\Cache\Memcached" instead.'), \E_USER_DEPRECATED);
 
-    /**
-     * Create a new cache object
-     * @param string $location Location string (from SimplePie::$cache_location)
-     * @param string $name     Unique ID for the cache
-     * @param string $type     Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
-     */
-    public function __construct($location, $name, $type) {
-        $this->options = array(
-            'host'   => '127.0.0.1',
-            'port'   => 11211,
-            'extras' => array(
-                'timeout' => 3600, // one hour
-                'prefix'  => 'simplepie_',
-            ),
-        );
-        $this->options = SimplePie_Misc::array_merge_recursive($this->options, SimplePie_Cache::parse_URL($location));
-
-        $this->name = $this->options['extras']['prefix'] . md5("$name:$type");
-
-        $this->cache = new Memcached();
-        $this->cache->addServer($this->options['host'], (int)$this->options['port']);
-    }
-
-    /**
-     * Save data to the cache
-     * @param array|SimplePie $data Data to store in the cache. If passed a SimplePie object, only cache the $data property
-     * @return bool Successfulness
-     */
-    public function save($data) {
-        if ($data instanceof SimplePie) {
-            $data = $data->data;
-        }
-
-        return $this->setData(serialize($data));
-    }
-
-    /**
-     * Retrieve the data saved to the cache
-     * @return array Data for SimplePie::$data
-     */
-    public function load() {
-        $data = $this->cache->get($this->name);
-
-        if ($data !== false) {
-            return unserialize($data);
-        }
-        return false;
-    }
-
-    /**
-     * Retrieve the last modified time for the cache
-     * @return int Timestamp
-     */
-    public function mtime() {
-        $data = $this->cache->get($this->name . '_mtime');
-        return (int) $data;
-    }
-
-    /**
-     * Set the last modified time to the current time
-     * @return bool Success status
-     */
-    public function touch() {
-        $data = $this->cache->get($this->name);
-        return $this->setData($data);
-    }
-
-    /**
-     * Remove the cache
-     * @return bool Success status
-     */
-    public function unlink() {
-        return $this->cache->delete($this->name, 0);
-    }
-
-    /**
-     * Set the last modified time and data to Memcached
-     * @return bool Success status
-     */
-    private function setData($data) {
-
-        if ($data !== false) {
-            $this->cache->set($this->name . '_mtime', time(), (int)$this->options['extras']['timeout']);
-            return $this->cache->set($this->name, $data, (int)$this->options['extras']['timeout']);
-        }
-
-        return false;
-    }
+if (\false) {
+	/** @deprecated since SimplePie 1.x, use "SimplePie\Cache\Memcached" instead */
+	class SimplePie_Cache_Memcached extends Memcached
+	{
+	}
 }
-
-class_alias('SimplePie_Cache_Memcached', 'SimplePie\Cache\Memcached', false);
